+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Fork 0.0.1 (1.3.2015)
 
+- improved testing and refactoring
+- various new field basic types
+- improved testing
+
 ### Fork 0.0.0 (17.2.2015)
 
 - first draft initialized with base project structure